Factor parameter bits out of NodeInvoker into ParameterManager.
